+++ client/src/user/student/hook/useCarousel.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState, useRef, useCallback } from "react";
 
 interface CarouselOptions<T> {
-  originalList: T[];
+  originalLists: T[];
   itemHeight: number;
   itemsPerPage: number;
   autoInterval?: number;
@@ -9,7 +9,7 @@ interface CarouselOptions<T> {
 }
 
 const useCarousel = <T>({
-  originalList,
+  originalLists,
   itemHeight,
   itemsPerPage,
   autoInterval = 3000,
@@ -22,11 +22,11 @@ const useCarousel = <T>({
   const createCarouselList = (): T[] => {
     const repeatedList = Array.from(
       { length: itemsPerPage },
-      () => originalList
+      () => originalLists
     ).flat();
 
     // 첫 번째 리스트로 자연스럽게 넘어가기 위해 요소 추가
-    const overflowItems = originalList.slice(0, itemsPerPage);
+    const overflowItems = originalLists.slice(0, itemsPerPage);
 
     return [...repeatedList, ...overflowItems];
   };
@@ -37,8 +37,8 @@ const useCarousel = <T>({
     if (intervalRef.current) return;
     intervalRef.current = setInterval(() => {
       setIndex((prev) => {
-        if (prev === originalList.length - 1) {
-          setIndex(originalList.length);
+        if (prev === originalLists.length - 1) {
+          setIndex(originalLists.length);
 
           // 마지막 transition이 끝난 후에 transition을 제거하고 첫 번째로 점프
           setTimeout(() => {
